replace NodeTypeSet with NodeTypePool
